improved script a little bit

+++ mara.sh
@@ -1,8 +1,15 @@
 #!/bin/sh
 
+echo "Clean build? Press 'y' to clean!"
+read -n1 c
+echo ""
+
+if [ $c == "y" ]
+then
 rm mara Makefile
 cmake .
 make
+fi
 
 startTime=`date +%Y-%m-%d_%H-%M-%S`
 
@@ -12,15 +19,16 @@ counterfile=testlogs/$startTime/running.txt
 mkdir testlogs
 mkdir $logpath
 touch $counterfile
+echo "0" >> $counterfile
 
 function maraTest {
   cf=$2
   n=$(tail -n 1 $cf)
   echo $((n+1)) >> $cf
+  echo "started mara with seed $1"
   ./mara test 50000 0.85 0.2 4 16 1000 1 $1 >> $logpath/$1.log
   n=$(tail -n 1 $cf)
   echo $((n-1)) >> $cf
-  echo "started mara with seed $1"
 }
 
 n=0
